Implementing custom highlighting in a simpler way to work for advanced field types.
