+++ graylog2-web-interface/src/views/components/highlighting/CustomHighlighting.test.tsx
