+++ graylog2-web-interface/src/views/components/highlighting/PossiblyHighlight.test.tsx
