+++ graylog2-web-interface/src/views/components/messagelist/SearchQueryHighlights.tsx
@@ -20,7 +20,7 @@ import PropTypes from 'prop-types';
 import { AdditionalContext } from 'views/logic/ActionContext';
 import { DEFAULT_HIGHLIGHT_COLOR } from 'views/Constants';
 
-import PossiblyHighlight from './PossiblyHighlight';
+import PossiblyHighlight from '../highlighting/PossiblyHighlight';
 
 type Props = {
   field: string,
